# docs/source/specs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "The API for Role Based Access Control.",
    "version": "1.0.0",
    "title": "Role Based Access Control",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://opensource.org/licenses/AGPL-3.0"
    }
  },
  "security": [
    {
      "basic_auth": []
    }
  ],
...
  "paths": {
...
    "/roles/{uuid}/": {
      "get": {
        "tags": [
          "Role"
        ],
        "summary": "Get a role in the tenant",
        "operationId": "getRole",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "ID of role to get",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/ScopeFilter"
          }
        ],
        "responses": {
          "200": {
            "description": "A role object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleWithAccess"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Insufficient permissions to get role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error403"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Role"
        ],
        "summary": "Delete a role in the tenant",
        "operationId": "deleteRole",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "ID of role to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Role deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Insufficient permissions to delete role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error403"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Role"
        ],
        "summary": "Update a role in the tenant",
        "operationId": "updateRole",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "ID of role to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePut"
              }
            }
          },
          "description": "Update to a role",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Role updated"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Insufficient permissions to update role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error403"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Role"
        ],
        "summary": "Patch a role in the tenant",
        "operationId": "patchRole",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "ID of role to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePatch"
              }
            }
          },
          "description": "Patch to a role"
        },
        "responses": {
          "200": {
            "description": "Role patched",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleWithAccess"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Insufficient permissions to patch role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error403"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://console.redhat.com/{basePath}",
      "description": "Production Server",
      "variables": {
        "basePath": {
          "default": "api/rbac/v1"
        }
      }
    },
    {
      "url": "https://console.stage.redhat.com/{basePath}",
      "description": "Staging Server",
      "variables": {
        "basePath": {
          "default": "api/rbac/v1"
        }
      }
    },
    {
      "url": "http://localhost:{port}/{basePath}",
      "description": "Development Server",
      "variables": {
        "port": {
          "default": "8000"
        },
        "basePath": {
          "default": "api/rbac/v1"
        }
      }
    }
  ],
  "components": {
    "parameters": {
...
      "ScopeFilter": {
        "in": "query",
        "name": "scope",
        "required": false,
        "description": "Parameter for filtering resource by scope.",
        "schema": {
          "type": "string",
          "enum": [
            "org_id",
            "principal"
          ],
          "default": "org_id"
        }
      },
...
    },
...
    "securitySchemes": {
      "basic_auth": {
        "type": "http",
        "description": "The userid/password is needed when accessing this API externally",
        "scheme": "basic"
      }
    },
    "schemas": {
      "Error": {
        "required": [
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string",
                  "example": "Not found."
                },
                "status": {
                  "type": "string",
                  "example": "403"
                }
              }
            }
          }
        }
      },
...
      "Error403": {
        "required": [
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string",
                  "example": "You do not have permission to perform this action."
                },
                "source": {
                  "type": "string",
                  "example": "detail"
                },
                "status": {
                  "type": "string",
                  "example": "403"
                }
              }
            }
          }
        }
      },
      "UUID": {
        "type": "object",
        "required": [
          "uuid"
        ],
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid",
            "example": "57e60f90-8c0c-4bd1-87a0-2143759aae1c"
          }
        }
      },
      "Timestamped": {
        "type": "object",
        "required": [
          "created",
          "modified"
        ],
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time",
            "example": "2019-01-21T17:32:28Z"
          },
          "modified": {
            "type": "string",
            "format": "date-time",
            "example": "2019-03-04T07:25:58Z"
          }
        }
      },
...
      "ResourceDefinitionFilter": {
          "oneOf": [
              {
                "$ref": "#/components/schemas/ResourceDefinitionFilterOperationEqual"
              },
              {
                "$ref": "#/components/schemas/ResourceDefinitionFilterOperationIn"
              }
          ]
     },
      "ResourceDefinitionFilterOperationEqual": {
        "type": "object",
        "required": [
            "key",
            "operation",
            "value"
        ],
        "properties": {
          "key": {
            "type": "string",
            "example": "cost-management.aws.account"
          },
          "operation": {
            "type": "string",
            "enum": [
              "equal"
            ]
          },
          "value": {
            "type": "string",
            "example": "123456",
            "nullable": true
          }
        }
      },
      "ResourceDefinitionFilterOperationIn": {
        "type": "object",
        "required": [
            "key",
            "operation",
            "value"
        ],
        "properties": {
          "key": {
            "type": "string",
            "example": "cost-management.aws.account"
          },
          "operation": {
            "type": "string",
            "enum": [
              "in"
            ]
          },
          "value": {
            "type": "array",
            "items": {
              "type": "string",
              "nullable": true
            },
            "example": ["value1", null, "value2"]
          }
        }
      },
      "ResourceDefinition": {
        "required": [
          "attributeFilter"
        ],
        "properties": {
          "attributeFilter": {
            "$ref": "#/components/schemas/ResourceDefinitionFilter"
          }
        }
      },
      "Access": {
        "required": [
          "permission",
          "resourceDefinitions"
        ],
        "properties": {
          "permission": {
            "type": "string",
            "example": "cost-management:*:read"
          },
          "resourceDefinitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceDefinition"
            }
          }
        }
      },
      "Role": {
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "RoleA"
          },
          "display_name": {
            "type": "string",
            "example": "ARoleName"
          },
          "description": {
            "type": "string",
            "example": "A description of RoleA"
          }
        }
      },
...
      "RoleOut": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Role"
          },
          {
            "$ref": "#/components/schemas/UUID"
          },
          {
            "$ref": "#/components/schemas/Timestamped"
          },
          {
            "properties": {
              "policyCount": {
                "type": "integer",
                "minimum": 0
              },
              "accessCount": {
                "type": "integer",
                "minimum": 0
              },
              "applications": {
                "type": "array",
                "items": {
                  "type": "string",
                  "example": "catalog"
                }
              },
              "system": {
                "type": "boolean",
                "default": false
              },
              "platform_default": {
                "type": "boolean",
                "default": false
              },
              "admin_default": {
                "type": "boolean",
                "default": false
              },
              "external_role_id": {
                "type": "string",
                "example": "ExternalRoleId"
              },
              "external_tenant": {
                "type": "string",
                "example": "ExternalTenant"
              }
            }
          }
        ]
      },
...
      "RolePatch": {
        "properties": {
          "name": {
            "type": "string",
            "example": "RoleA"
          },
          "display_name": {
            "type": "string",
            "example": "ARoleName"
          },
          "description": {
            "type": "string",
            "example": "A description of RoleA"
          }
        }
      },
      "RolePut": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RolePatch"
          },
          {
            "type": "object",
            "required": [
              "access"
            ],
            "properties": {
              "access": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Access"
                },
                "description": "List of access permissions. This array can be empty."
              }
            }
          }
        ]
      },
      "RoleWithAccess": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RoleOut"
          },
          {
            "type": "object",
            "required": [
              "access"
            ],
            "properties": {
              "access": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Access"
                }
              }
            }
          }
        ]
      },
...
    }
  }
}